requirements.txt
torch
transformers
ffmpeg-python
pysrt
sentencepiece